# server/docs/swagger.yaml
definitions:
  objects.AccountDTO:
    properties:
      login:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
...
info:
  contact: {}
  description: API for culinary recipes (BMSTU Web development cource project)
  title: Recipes API
  version: "0.1"
paths:
  /accounts/login:
    post:
      parameters:
      - description: Authentication data
        in: body
        name: account
        schema:
          $ref: '#/definitions/objects.AccountDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/objects.AccountDTO'
      summary: User's authorization
      tags:
      - Accounts
...
swagger: "2.0"